Added template support
This patch adds support for templates. Templates are a property on Objects
that specify an external file to load object properties by default. In
this implementation, templates are not visible to the end user, their
behaviour happens "automatically".

As it parses, the Map will generate a list of templates, with each being
identified by their ID.

Templates introduce the concept of a tile that is bound to a tileset
that is not part of the main map's list of tilesets. For this reason,
the tileset_index() function now returns Option<usize>. Internally, an
enum is used to signify whether the tileset is map or template bound.

I had to change the signature of the Resource cache, because the
generator function now needs access to the cache.