Add template support (#170)
* Added template support

This patch adds support for templates. Templates are a property on Objects
that specify an external file to load object properties by default. In
this implementation, templates are not visible to the end user, their
behaviour happens "automatically".

As it parses, the Map will generate a list of templates, with each being
identified by their ID.

Templates introduce the concept of a tile that is bound to a tileset
that is not part of the main map's list of tilesets. For this reason,
the tileset_index() function now returns Option<usize>. Internally, an
enum is used to signify whether the tileset is map or template bound.

I had to change the signature of the Resource cache, because the
generator function now needs access to the cache.

* Handle error on template base_path.parent()

* LayerTileData object now can contain a templ. arc

Rather than an id into some list, this now contains an arc to the
tileset directly. Makes the code cleaner, and allows caching to behave
better, but can have performance implications.

* Removed some more template list Vecs

These should have been removed last commit, whoops!

* Responded to review comments

* Don't bind attributes property to a name

Squishes an "unused" warning.

* Cargo fmt and updated test

* Update changelog

* Clippy fixes

* Simplify LayerTileData and create ObjectTileData

* Fix clippy warnings

* Simplify implementation

* Apply suggestions from code review

* Remove outdated comment

* Address PR comments

* Trigger CI & handle review comment

Co-authored-by: alexdevteam <alexpro820@gmail.com>
Co-authored-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>
Co-authored-by: aleokdev <aleok.inf@gmail.com>