Fix CI
CI was failing with an error related to include! that it was not doing before.
This looks like a regression on someone else's part to be honest.
I'm not sure what we can do apart from changing the edition or just
skipping these tests, which is what I've done here.